users/backend/requirements.txt
fastapi
sqlalchemy
pydantic
python-dotenv
httpx
starlette
passlib[bcrypt]
python-jose
uvicorn
psycopg2-binary
python-multipart